Refactor: Extract OnThisPage component and logic
